refactor :: Country 저장시 Enum 활용
