+++ motung-i-api/src/main/kotlin/kr/motung_i/backend/domain/tour/formatter/TourFormatter.kt
@@ -2,7 +2,7 @@ package kr.motung_i.backend.domain.tour.formatter
 
 import kr.motung_i.backend.global.geojson.dto.District
 import kr.motung_i.backend.global.geojson.dto.Region
-import kr.motung_i.backend.global.geojson.enums.Country
+import kr.motung_i.backend.persistence.tour.entity.Country
 
 interface TourFormatter {
     fun formatToTourFilterRegion(region: Region, country: Country): String
