+++ motung-i-api/src/main/kotlin/kr/motung_i/backend/domain/tour/formatter/impl/KoreaTourFormatterImpl.kt
@@ -3,7 +3,7 @@ package kr.motung_i.backend.domain.tour.formatter.impl
 import kr.motung_i.backend.domain.tour.formatter.TourFormatter
 import kr.motung_i.backend.global.geojson.dto.District
 import kr.motung_i.backend.global.geojson.dto.Region
-import kr.motung_i.backend.global.geojson.enums.Country
+import kr.motung_i.backend.persistence.tour.entity.Country
 
 class KoreaTourFormatterImpl : TourFormatter {
     override fun formatToTourFilterRegion(region: Region, country: Country): String {
