+++ motung-i-api/src/main/kotlin/kr/motung_i/backend/global/geojson/LocalsFactory.kt
@@ -8,7 +8,7 @@ import kr.motung_i.backend.global.geojson.dto.District
 import kr.motung_i.backend.global.geojson.dto.Local
 import kr.motung_i.backend.global.geojson.dto.Neighborhood
 import kr.motung_i.backend.global.geojson.dto.Region
-import kr.motung_i.backend.global.geojson.enums.Country
+import kr.motung_i.backend.persistence.tour.entity.Country
 import kr.motung_i.backend.global.geojson.formatter.LocalFormatterService
 import org.geolatte.geom.*
 import org.geolatte.geom.crs.CoordinateReferenceSystems
