+++ motung-i-api/src/main/kotlin/kr/motung_i/backend/global/geojson/dto/Local.kt
@@ -1,6 +1,6 @@
 package kr.motung_i.backend.global.geojson.dto
 
-import kr.motung_i.backend.global.geojson.enums.Country
+import kr.motung_i.backend.persistence.tour.entity.Country
 
 data class Local(
     val country: Country,
