+++ motung-i-domain/src/main/kotlin/kr/motung_i/backend/persistence/tour/entity/Local.kt
@@ -2,6 +2,8 @@ package kr.motung_i.backend.persistence.tour.entity
 
 import jakarta.persistence.Column
 import jakarta.persistence.Embeddable
+import jakarta.persistence.EnumType
+import jakarta.persistence.Enumerated
 
 @Embeddable
 data class Local(
@@ -9,7 +11,8 @@ data class Local(
     val localAlias: String,
 
     @Column(nullable = false)
-    val country: String,
+    @Enumerated(EnumType.STRING)
+    val country: Country,
 
     @Column(nullable = false)
     val regionAlias: String,
